+++ test/clojure/nrepl/core_test.clj
@@ -86,7 +86,19 @@
 
 (defn- strict-transport? []
   ;; TODO: add transit here.
-  (= *transport-fn* #'transport/edn))
+  (or (= *transport-fn* #'transport/edn)
+      (when-require 'fastlane.core
+                    (or (= *transport-fn* #'fastlane.core/transit+msgpack)
+                        (= *transport-fn* #'fastlane.core/transit+json)
+                        (= *transport-fn* #'fastlane.core/transit+json-verbose)))))
+
+(defn- check-response-format
+  "checks response against spec, if available it to do a spec check later"
+  [resp]
+  (when-require 'nrepl.spec
+                (when-not (#'clojure.spec.alpha/valid? :nrepl.spec/message resp)
+                  (throw (Exception. (#'clojure.spec.alpha/explain-str :nrepl.spec/message resp)))))
+  resp)
 
 (defn- clean-response
   "Cleans a response to help testing.
@@ -116,7 +128,8 @@
     (cond-> resp
       true                      de-identify
       (not (strict-transport?)) normalize-status
-      (not (strict-transport?)) keywordize-truncated-keys)))
+      (not (strict-transport?)) keywordize-truncated-keys
+      (strict-transport?)       check-response-format)))
 
 (def-repl-test eval-literals
   (are [literal] (= (binding [*ns* (find-ns 'user)] ; needed for the ::keyword
@@ -191,7 +204,7 @@
 
 (def-repl-test session-lifecycle
   (is (= #{:error :unknown-session :done}
-         (-> (message timeout-client {:session "abc"})
+         (-> (message timeout-client {:session "00000000-0000-0000-0000-000000000000"})
              combine-responses
              clean-response
              :status)))